Update Assignment4.java
I guess this could be the final version, any changes I made were basically commentary. Also I changed input to System.in.
Do you think we should have a map in this program: on the assg in Moodle it says "A short problem that uses maps, lists, arrays, and a permutation." But a hash set could sub for a map, maybe...
I just don't see where we could realistically apply the map that it would make a big difference.
You can turn it in if you want,or tell me to do it.